rename and simplify tree
